requirements.txt
altair==5.2.0
annotated-types==0.6.0
appdirs==1.4.4
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bqplot==0.12.43
branca==0.7.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
colour==0.1.5
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dateutils==0.6.12
decorator==5.1.1
duckdb==0.9.2
entrypoints==0.4
entsoe-py==0.6.7
et-xmlfile==1.1.0
executing==2.0.1
extra-streamlit-components==0.1.70
Faker==24.0.0
favicon==0.7.0
filelock==3.13.1
fiona==1.9.5
folium==0.16.0
fonttools==4.49.0
forex-python==1.8
gdown==5.1.0
geojson==3.1.0
geopandas==0.14.3
GHEtool==2.2.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
gw_dsl_parser==0.1.44
htbuilder==0.6.2
idna==3.6
importlib-metadata==7.0.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.18.2
ipython==8.22.2
ipytree==0.2.2
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.10
kanaries_track==0.0.4
kiwisolver==1.4.5
leafmap==0.31.5
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
numpy==1.26.4
numpy-financial==1.0.0
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
plotly==5.19.0
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
pyecharts==2.0.5
pygfunction==2.2.2
Pygments==2.17.2
pygwalker==0.4.6
PyJWT==2.8.0
pymdown-extensions==10.7.1
pyparsing==3.1.1
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
pystac==1.9.0
pystac-client==0.7.6
python-box==7.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
scipy==1.12.0
scooby==0.9.2
SecondaryCoolantProps==1.1
segment-analytics-python==2.2.3
shapely==2.0.3
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.27
sqlglot==22.2.0
st-annotated-text==4.0.1
stack-data==0.6.3
streamlit==1.31.1
streamlit-aggrid==0.3.4.post3
streamlit-authenticator==0.3.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-echarts==0.4.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-folium==0.18.0
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.3
streamlit-lottie==0.0.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
traittypes==0.2.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
wasmtime==12.0.0
watchdog==4.0.0
wcwidth==0.2.13
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==4.0.10
xyzservices==2023.10.1
zipp==3.17.0
